DEX-2053 create pm UI module (#416)
